refactor(cli): fold streamArtifactProjection into its owner and drop unreachable guards
- delete streamArtifactProjection.ts; its types and projection live with the subscription that owns them
- totalize COMPACT_TODO_ROW_PRIORITY and drop the unreachable default arm plus assertNever
- TranscriptReader drops the React 19-inert isMounted latch and the spillPathsKey round trip
